X out force_regular + only light stemcells spec [#140953633]
Our acceptance tests integrate with the production bosh.io API

When this test was written, the API returned AWS Xen HVM stemcells
that had *only light versions*. This meant that when you did
`force_regular`, the check should return no stemcells.

This isn't the case anymore and we can't rely on the state of bosh.io
ever allowing this test to pass again in the future, so I am commenting
it out.

[finishes #140953633]