Add support to download stemcells from s3/gcs with auth
Work is currently being done to start building FIPS stemcells. These stemcells
will not be publicly available, but will still be discoverable via bosh.io.

These changes will allow access and secret keys to be passed as part of the source
configuration. When fetching the stemcell, it will use these credentials to build a
minio client that can download the stemcell.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>